Move relvance related functions
